Fix uint overflow in premium_unti parameter
PremUntil contains a timestamp of the day when the premium is over. That's a number bigger than uint_32 limit, thus it was overflowing and returning wrong value. This commit updates the protobuf, the structures and methods to uint64.

+++ src/api/models/models_test.go
@@ -36,7 +36,7 @@ func createCharacterMessage(sex uint32) *login_proto_messages.Character {
 func createSessionMessage(isPremium bool) *login_proto_messages.Session {
 	return &login_proto_messages.Session{
 		IsPremium:    isPremium,
-		PremiumUntil: defaultNumber,
+		PremiumUntil: uint64(defaultNumber),
 		SessionKey:   defaultString,
 		LastLogin:    defaultNumber,
 	}
